feature: addition of data and pyspark_bike code
