+++ app/application.py
@@ -1,4 +1,4 @@
-from common.ldi_logger import LdiLogger
+from common_ldi.ldi_logger import LdiLogger
 from pyspark.sql import DataFrame
 
 def main() -> None:
