+++ requirements.txt
@@ -1,5 +1,45 @@
+beautifulsoup4==4.12.3
+cachetools==5.5.0
+certifi==2024.8.30
+charset-normalizer==3.4.0
+chispa==0.10.1
+exceptiongroup==1.2.2
+google==3.0.0
+google-api-core==2.23.0
+google-api-python-client==2.153.0
+google-auth==2.36.0
+google-auth-httplib2==0.2.0
+google-auth-oauthlib==1.2.1
+googleapis-common-protos==1.66.0
+grpcio==1.67.1
+grpcio-status==1.67.1
+httplib2==0.22.0
+idna==3.10
+iniconfig==2.0.0
+numpy==2.0.2
+oauthlib==3.2.2
+packaging==24.2
+pandas==2.2.3
+pluggy==1.5.0
+prettytable==3.12.0
+proto-plus==1.25.0
+protobuf==5.28.3
 py4j==0.10.9.7
+pyarrow==18.0.0
+pyasn1==0.6.1
+pyasn1_modules==0.4.1
+pyparsing==3.2.0
 pyspark==3.5.3
-
-pytest~=8.3.3
-chispa~=0.10.1
+pytest==8.3.3
+python-dateutil==2.9.0.post0
+pytz==2024.2
+requests==2.32.3
+requests-oauthlib==2.0.0
+rsa==4.9
+six==1.16.0
+soupsieve==2.6
+tomli==2.1.0
+tzdata==2024.2
+uritemplate==4.1.1
+urllib3==2.2.3
+wcwidth==0.2.13
